Monday Dev

+++ resources/views/cp/advanced/roles/index.blade.php
@@ -21,13 +21,15 @@
 			<div class="{{ $mainSmCols }} {{ $mainMdCols }} {{ $mainLgCols }} main">
 				@include('cp._partials.message')
 				@include('cp._partials.pageTitle')
-				<div class="row">
-					<div class="col">
-						<ul class="list-unstyled list-inline buttons">
-							<li class="list-inline-item"><a href="/cp/advanced/roles/create" title="Add Role" class="btn btn-outline-success"><i class="icon fa fa-plus" aria-hidden="true"></i>Add Role</a></li>
-						</ul>
+				@if ($currentUser->hasPermission('create_roles'))
+					<div class="row">
+						<div class="col">
+							<ul class="list-unstyled list-inline buttons">
+								<li class="list-inline-item"><a href="/cp/advanced/roles/create" title="Add Role" class="btn btn-outline-success"><i class="icon fa fa-plus" aria-hidden="true"></i>Add Role</a></li>
+							</ul>
+						</div>
 					</div>
-				</div>
+				@endif
 				<div class="content padding bg-white">	
 					<table id="datatable" class="table table-striped table-bordered table-hover" cellspacing="0" border="0" cellpadding="0" width="100%">
 						<thead>
@@ -42,13 +44,21 @@
 								<tr>
 									<td>{{ $role->id }}</td>
 									<td>{{ $role->title }}</td>
-									<td class="actions dropdown text-center" id="submenu">
-										<a href="javascript:void(0);" title="Role Actions" class="dropdown-toggle" id="pageActions" data-toggle="dropdown"><i class="fa fa-ellipsis-v" aria-hidden="true"></i></a>
-										<ul class="dropdown-menu dropdown-menu-right">
-											<li class="dropdown-item gf-info"><a href="/cp/advanced/roles/{{ $role->id }}/edit" title="Edit Role"><i class="icon fa fa-pencil" aria-hidden="true"></i>Edit Role</a></li>
-											<li class="dropdown-item gf-danger"><a href="/cp/advanced/roles/{{ $role->id }}/delete" title="Delete Role"><i class="icon fa fa-trash" aria-hidden="true"></i>Delete Role</a></li>
-										</ul>
-									</td>
+									@if ($currentUser->hasPermission('edit_roles') || $currentUser->hasPermission('delete_roles'))
+										<td class="actions dropdown text-center" id="submenu">
+											<a href="javascript:void(0);" title="Role Actions" class="dropdown-toggle" id="pageActions" data-toggle="dropdown"><i class="fa fa-ellipsis-v" aria-hidden="true"></i></a>
+											<ul class="dropdown-menu dropdown-menu-right">
+												@if ($currentUser->hasPermission('edit_roles'))
+													<li class="dropdown-item gf-info"><a href="/cp/advanced/roles/{{ $role->id }}/edit" title="Edit Role"><i class="icon fa fa-pencil" aria-hidden="true"></i>Edit Role</a></li>
+												@endif
+												@if ($currentUser->hasPermission('delete_roles'))
+													<li class="dropdown-item gf-danger"><a href="/cp/advanced/roles/{{ $role->id }}/delete" title="Delete Role"><i class="icon fa fa-trash" aria-hidden="true"></i>Delete Role</a></li>
+												@endif
+											</ul>
+										</td>
+									@else
+										<td>&nbsp;</td>
+									@endif
 								</tr>
 							@endforeach
 						</tbody>
