+++ resources/views/cp/locations/edit.blade.php
@@ -167,7 +167,7 @@
 								<a href="/cp/locations" title="Cancel" class="btn btn-outline-secondary cancel-button" tabindex="16" title="Cancel">Cancel</a>
 							@endif
 							<button type="submit" name="submit" id="submit" class="btn btn-outline-primary" tabindex="15" title="Save Changes">Save Changes</button>
-							@if ($currentUser->hasPermission('delete_locations'))
+							@if ($currentUser->hasPermission('delete_locations') && !in_array($location->id, $defaultLocationIds))
 								<a href="/cp/locations/{{ $location->id }}/delete" title="Delete Location" class="pull-right btn btn-outline-danger">Delete Location</a>
 							@endif
 						</div>
